Extract the method for plugging in named path segments
